% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper0.R
\name{calc_distance}
\alias{calc_distance}
\title{calc_distance".}
\usage{
calc_distance(
  file_tsv = NULL,
  file_sample_location = NULL,
  runnames = NULL,
  omit_popid = NULL,
  amp_euc = 1,
  amp_bin = 0,
  param = param0
)
}
\description{
calc_distance".
}
