% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper1_proto0.R
\name{plot.popmap}
\alias{plot.popmap}
\title{plot.popmap".}
\usage{
\method{plot}{popmap}(
  popmap,
  param = param0,
  perm = 0,
  resp = NULL,
  main = NULL,
  label_gid = FALSE
)
}
\description{
plot.popmap".
}
