% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper1_proto0.R
\name{plot_popmap}
\alias{plot_popmap}
\title{plot_popmap".}
\usage{
plot_popmap(
  popmap,
  param = param0,
  perm = 0,
  resp = NULL,
  main = NULL,
  labels = FALSE
)
}
\description{
plot_popmap".
}
