% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper1_proto0.R
\name{popmap.plot}
\alias{popmap.plot}
\title{popmap.plot".}
\usage{
popmap.plot(
  popmap,
  param = param0,
  perm = 0,
  resp = NULL,
  main = NULL,
  labels = NULL,
  sample_group_name = "",
  peak = TRUE,
  power = TRUE,
  main_add = FALSE,
  cid_sm = popmap$cid_sm,
  pals = popmap$pals
)
}
\description{
popmap.plot".
}
