% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper1_proto0.R
\name{popmap.plotdata}
\alias{popmap.plotdata}
\title{popmap.plotdata".}
\usage{
popmap.plotdata(
  pcoa,
  data,
  param = param0,
  sample_group_name = NULL,
  flag_lim_out = 0,
  plot_mask = NULL,
  col1 = "red",
  col2 = "gray",
  tagid = NULL
)
}
\description{
popmap.plotdata".
}
