% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper0g.R
\name{popmap.readdata}
\alias{popmap.readdata}
\title{popmap.readdata".}
\usage{
popmap.readdata(
  file_tsv = NULL,
  file_sample_location = NULL,
  omit_popid = NULL,
  amp_euc = 1,
  amp_bin = 0,
  param = param0
)
}
\description{
popmap.readdata".
}
