% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmapper0.R
\docType{package}
\name{popmapper-package}
\alias{popmapper-package}
\alias{_PACKAGE}
\alias{popmapper}
\title{Experimental R package.}
\description{
Experimental R package.
}
\author{
\strong{Maintainer}: Hiroshi C. Ito \email{hiroshibeetle@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
